refactor: reorganize files for app usage
