+++ internal/controller/user.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/gin-gonic/gin"
 	jsoniter "github.com/json-iterator/go"
-	"github.com/julianstephens/license-server/model"
-	"github.com/julianstephens/license-server/pkg/httputil"
+	"github.com/julianstephens/license-server/internal/controller/httputil"
+	"github.com/julianstephens/license-server/internal/model"
 	"github.com/julianstephens/license-server/service"
 )
