+++ internal/licensemanager/manager.go
@@ -1,0 +1,6 @@
+package licensemanager
+
+func CreateSoftwareKeyPair() {}
+func GenerateLicense()       {}
+func ExportLicense()         {}
+func ValidateLicense()       {}
