+++ internal/model/base.go
