+++ internal/model/licenses.go
