+++ internal/model/products.go
