+++ internal/model/rules.go
