+++ internal/router/router.go
@@ -9,10 +9,10 @@ import (
 	"github.com/chenyahui/gin-cache/persist"
 	"github.com/gin-contrib/gzip"
 	"github.com/gin-gonic/gin"
-	"github.com/julianstephens/license-server/controller"
 	docs "github.com/julianstephens/license-server/docs"
+	"github.com/julianstephens/license-server/internal/controller"
+	"github.com/julianstephens/license-server/internal/middleware"
 	"github.com/julianstephens/license-server/pkg/logger"
-	"github.com/julianstephens/license-server/pkg/middleware"
 	sloggin "github.com/samber/slog-gin"
 	"gorm.io/gorm"
 )
