+++ scripts/atlas-gorm-loader.go
@@ -7,7 +7,7 @@ import (
 
 	_ "ariga.io/atlas-go-sdk/recordriver"
 	"ariga.io/atlas-provider-gorm/gormschema"
-	"github.com/julianstephens/license-server/model"
+	"github.com/julianstephens/license-server/internal/model"
 )
 
 // Define the models to generate migrations for.
